Move Kubevirt code to frontend/public/kubevirt
with intention to modify OKD codebase when absolutely neccessary only.

This approach enables straightforward merging of upstream changes
to our fork.

On the otherhand, the structure is still good-enough to allow smooth
merging of kubevirt to upstream in the future.
