+++ frontend/__tests__/kubevirt/vm.tsx
@@ -1,0 +1,5 @@
+describe('Kubevirt', () => {
+  it('VM renders', () => {
+    expect(1).toBe(1); // TODO
+  });
+});
