+++ frontend/public/kubevirt/_style.scss
@@ -1,0 +1,1 @@
+@import "components/vm";
