+++ frontend/public/kubevirt/components/_vm.scss
