+++ frontend/public/kubevirt/components/factory/okdfactory.tsx
@@ -1,0 +1,3 @@
+// indirect reuse of OKD in kubevirt
+
+export * from '../../../components/factory';
