+++ frontend/public/kubevirt/components/utils/okdutils.tsx
@@ -1,0 +1,5 @@
+// indirect reuse of OKD in kubevirt
+
+export * from '../../../components/utils';
+export { breadcrumbsForOwnerRefs } from '../../../components/utils/breadcrumbs';
+
