+++ frontend/public/kubevirt/models/index.ts
@@ -1,0 +1,4 @@
+// ./vm.ts already imported by okd-core models due to feature detection
+
+export * from '../../models'; 
+
