+++ frontend/public/models/index.ts
@@ -735,6 +735,8 @@ export const ServiceBindingModel: K8sKind = {
   id: 'servicebinding'
 };
 
+export * from '../kubevirt/models/vm'; // needed when setting features and resource pages
+
 export const LimitRangeModel: K8sKind = {
   label: 'Limit Range',
   apiVersion: 'v1',
@@ -745,44 +747,5 @@ export const LimitRangeModel: K8sKind = {
   kind: 'LimitRange',
   id: 'limitrange',
   labelPlural: 'Limit Ranges'
-};
-
-export const VirtualMachineModel: K8sKind = {
-  label: 'Virtual Machine',
-  labelPlural: 'Virtual Machines',
-  apiVersion: 'v1alpha2',
-  path: 'virtualmachines',
-  apiGroup: 'kubevirt.io',
-  plural: 'virtualmachines',
-  abbr: 'VM',
-  namespaced: true,
-  kind: 'VirtualMachine',
-  id: 'virtualmachine'
-};
-
-export const VirtualMachineInstanceModel: K8sKind = {
-  label: 'Virtual Machine Instance',
-  labelPlural: 'Virtual Machine Instances',
-  apiVersion: 'v1alpha2',
-  path: 'virtualmachineinstances',
-  apiGroup: 'kubevirt.io',
-  plural: 'virtualmachineinstances',
-  abbr: 'VMI',
-  namespaced: true,
-  kind: 'VirtualMachineInstance',
-  id: 'virtualmachineinstance'
-};
-
-export const VirtualMachineInstancePresetModel: K8sKind = {
-  label: 'Virtual Machine Instance Preset',
-  labelPlural: 'Virtual Machine Instance Presets',
-  apiVersion: 'v1alpha2',
-  path: 'virtualmachineinstancepresets',
-  apiGroup: 'kubevirt.io',
-  plural: 'virtualmachineinstancepresets',
-  abbr: 'VMIP',
-  namespaced: true,
-  kind: 'VirtualMachineInstancePreset',
-  id: 'virtualmachineinstancepreset'
 };
