+++ frontend/public/style.scss
@@ -92,5 +92,6 @@
 @import "components/overview";
 @import "components/project-overview";
 @import "components/list-view";
-@import "components/vm";
+
+@import "kubevirt/style";
